ALTER TRUNCATE and Correction
ALTER and TRUNCATE commands are now fully functional with the
filesystem. Also, the bug mentioned in the last commit was not due to
modifyTable in filesystem but rather a design flaw in BiblioBaseDBMS
which set records ID values to null rathar than actually deleting any
records (this was meant to preserve primary key data); this problem was
addressed by actually deleting records from a table and by storing the
max primary key value in a table which will in the future be taken from
a stored value in the table entree within the filesystem.