Commit Weirdness
The last commit didn't recognize these changes until after it was
synched. To reiterate the commit: ALTER commands now allow DROP
columnName, RENAME column or table and ADD a column or columns. The bug
mentioned two commits ago was not a bug in filesystem but was due to the
Table class setting all primary keys to null when "deleting" records
(which has been addressed).