Admin, Cat, Bug fix
Added pseudo-admin login which allows for the creation and deletion of
databases. Added the creation of a cat table which displays all other
tables in a database; created preventative measures against their
deletion (unless database is deleted). Fixed a bug whereby the creation
of multiple empty tables would create a run-on line (tables should be
separated by newlines).