Sysdatabases, Admin
Added an admin account which allows for database dropping, creation, or
name alteration. Added a sysdatabases table which stores all added,
modified or removed databases. The sysdatabase table only exists in the
admin folder. Unlike the cat table (which is in each database),
sysdatabase can only be viewed and modified in by the admin (login with
"admin").